단건조회완

+++ src/main/java/com/sparta/todos/repository/TodoRepository.java
@@ -3,5 +3,5 @@ package com.sparta.todos.repository;
 import com.sparta.todos.entity.Todo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TodoRepository extends JpaRepository<Todo, Integer> {
+public interface TodoRepository extends JpaRepository<Todo, Long> {
 }
